\name{stencila-package}
\alias{stencila-package}
\alias{stencila}
\docType{package}
\title{
Stencila
}
\description{
}
\details{
\tabular{ll}{
Package: \tab stencila\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-24\cr
License: \tab ISC\cr
}
}
\author{
Nokome Bentley <nokome.bentley@stenci.la>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
